Moved all 0mq things into a single header
